+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### WIP
 - Fix potential issues in test tasks
 - Add `normalize.css` to available modules to add
+- Fix bug in `addIgnored`
 
 #### 0.4.1 (November 5, 2013)
 - Update EditorConfig